Changes to make h5py-based tests work on Windows
Two separate issues are preventing h5py tests from working
on Windows.

The first is still somewhat vaguely understood, but I feel
fairly confident that this is a workable, non-hackish
solution.  Windows, unlike linux, permits multiple build
configurations (debug, release, etc) to exist in the same
build tree. When tests are built by CMake, they are built
for all enabled configs and can run uncondtionally. However,
when they use a custom COMMAND, as these h5py tests do, they
are restricted to running only for specifically enabled
configurations.

In common_build.cmake, we run ctest for our unit tests without
specifying any configuration, and so none of the h5py tests
were enabled. The fix here (hopefully) is just to add a
configuration specification to the ctest command.

The second is more straightforward. Again, because of this
multiconfig capability on Windows, the dakota binary is
placed in a folder named for the build configuration. The
h5py tests, which depend on this binary, assumed it was
located in src/ in the build tree. I am modifying the tests
to accept the Dakota path as a command line argument and
the macro that adds the tests to provide it.