DAK-2754: Remove existing HDF5 files
A user reported that Dakota sometimes crashes when attempting
to create an HDF5 output file after a crash, and that it's
necessary to delete the old HDF5 file before Dakota can run.

A possible/likely explanation is that Dakota was killed
(SIGKILL) or died abruptly without its signal handler being
called, and consequently its HDF5 file was left in an
unreadable or corrupt state. Dakota attempts to create/access
HDF5 files in H5F_ACC_TRUNK mode. In this mode, HDF5 tries
to open existing files with RDWR access. I have verified
that this fails when the file is corrupt (I generated a
corrupt file by SIGKILLing a previous Dakota run).

The fix is to remove existing HDF5 files before creating a
new one.