DAK-2732:  Add TimeoutAbsolute spec for slowly running tests on Mac
These EXPERIMENTAL tests are running slowly on the Jenkins Mac build
host and often lead to FAILURES which prevent the Nightly processes
from promoting devel to master.

Since these tests are EXPERIMENTAL, it seems natural to simply
workaround the issue by increasing the allowed running time by
adding:
near the top of the test file.  Not certain this will work, but it is
worth a try.  Again, noting the EXPERIMENTAL Label for these tests.

Request review by Brian Adams.