DAK-2732:  Fix stale output failure PARALLEL RHEL7
The serial tests consistently pass on RHEL7 but those have a fairly
high TimeoutDelay setting.
However, the TimeoutDelay setting had not been specified for the
PARALLEL tests, and occasional stale output failures have been occuring.

This commit replicates the TimeoutDelay setting so the increased limit
is allowed in PARALLEL tests.