Clean up of a fair amount of gymnastics in managing initial guesses for the MAP pre-solve and the Markov chain.
With the introduction of the inverse mapping in the ctor of
ProbabilityTransformModel, valid u-space initial guesses are available.
Bless a number of DIFFs for standardized_space tests in
dakota_rosenbrock_queso.in.

Kathryn should check to verify that code exhivits the desired warm-starting
for the high-to-low workflow.  Other cases are improved from before.