Pass keyword arguments down to mpirun invocation
tile_run_{dynamic,static} now accept **kwargs, which they
pass through to the Popen constructor that invokes mpirun. This
allows clients to specify, e.g., a cwd, streams for stdout/
stderr/stdin, and the like; whatever subprocess.Popen accepts
for their version of Python. By default, the stdout, stderr,
and stdin passed to Popen are those streams for the process.

It's a change that was suggested by a user a while ago, and
suddenly became helpful to me while working on a project.