Clarify volume_boxsize_limit convergence message
The old convergence message suggested that Dakota
interprets the volume_boxsize_limit keyword as a fraction
that the original volume was reduced, rather than as a
percentage of original volume. After a user complained about
the inconsistency between the message and Dakota's RefMan keyword
documentation, I reviewed the library documentation and the code
and determined that Dakota's RefMan probably is correct and
updated the message.