Initial specification of MUQ in the Dakota input file under bayes_calibration.
Currently, it is simply specified as "bayes_calibration muq".
The options exposed are mcmc type (dram or metropolis_hastings) and
chain_samples.  We can add more to the spec but this should be enough to
start testing now.  There is the start of a test in dakota_bayes_muq.in.