DP-47: Deploy GP surrogates in Dakota
 - add a 2-variable unit test

This mimics the setup in the surrogates unit test and uses the "on-demand"
API for constructing the GP.  The surrogate is evaluated sequentially
for the same four parameter values, and the responses compared to the
same gold values.  For reasons unknown (to me), the tolerance has to
be relaxed from 5.0e-7 to 3.0e-3 to pass.  Maybe related to a random number
seed?  Request review by @dtseidl .

Ideally, the same construction points can be obtained from the appropriate
model (herbie ?) brokered by Dakota along the lines of DP-57.