DP-65 : multi-index WIP needed for multivariate polynomials
 - add needed utils with unit tests adapted from Pecos
 - construct unit tests to produce corresponding behavior/values in Pecos::util

Note: A few utilities use two template types to allow Eigen::Map's of
objects to work alongside the mapped object type.

A bit more remains to be done before we can actually construct the
multi-variate polymials, but this is getting closer.  I'm committing
this state to avoid conflicts with @emridgw .