DP-118: Add support for batch evals to dakota.interfacing
Create new classes BatchParameters and BatchResults, which are
indexable and iterable containers of Parameters and Results
objects.

The return type of the module free function
dakota.interfacing.read_parameters_files() now depends on
whether the parameters file is for a batch evaluation or ordinary
sequential evaluation. In the batch case, it returns a
(BatchParameters, BatchResults) tuple. For a sequential eval,
it returns a (Parameters, Results) tuple. This behavior preserves
backward compatibility. It also means that in carefully written
code, it may be necessary to check the type.

Also, update the User's Manual documntation for
dakota.interfacing and its unit tests.