Surrogates: Base class and configOptions use in GP
DP:119: Created a Surrogate base class that contains
a pure virtual value function and virtual gradient and
hessian functions that print an error message when they
not implemented in a derived class.

Member variables in this base class include numQOI,
numSamples, and numVariables, as well as default and
working (i.e. actually used in surrogate construction)
ParameterLists.

DP:93 Use the ParameterList member function
validateParametersAndSetDefaults to check the ParameterList
input to the GP constructor and set missing values to default
values.

Started reorganizing the GP according to DP best practices.
Documentation is not yet complete.