Surrogates: Base class API in PolynomialRegression
DP-119: PolynomialRegression inherits from Surrogate, but
is still compatible with the original API.

DP-93: configOptions validation and default setting is now done
at the beginning of each build(...) call so that it is not possible
to build a surrogate with an invalid parameterList. This also enables
one to change configOptions through the setter and safely create a new
surrogate through the public build call.

Added a utility to convert string solver type names to solver types
based on the related function in the DataScaler.

PolynomialRegressionTest covers the validated and set defaults
optionsList case.