Surrogates: 3 ways to build using options and data
DP-119: defaultConfigOptions is now a pure virtual function. Each
surrogate has no-argument constructor that will instantiate the class
with the default config options in addition to a constructor that takes
a ParameterList and sets it to be the member variable configOptions but
does not build the surrogate model. The third constructor takes a
ParameterList, samples, and response matrices and builds the surrogate.

Unit test coverage for all three approaches is present for both the GP
and polynomial surrogates.