Surrogates: Derivatives for the polynomial model
DP-91: gradient and hessian member functions implemented
in the PolynomialRegression class with unit test coverage.

Fixed a bug in the NoScaler variant of the DataScaler class that was
causing strange behavior in the PolynomialRegressionTest.
The offset and scaling coefficients were not being set to 0 and 1,
respectively.

Comment cleanup and documentation are still needed.