Surrogates: best practices and documentation
DP-94 and DP-124: Roughly 50% done for the surrogates module.

DP-94: Reordered functions and reformatted code to adhere to DP
best practices. Renamed SurrogatesTools to surrogates_tools.

DP-124: Checked and updated existing doxygen documentation for the GP
files and wrote new documentation for surrogates_tools. There was no
existing documentation for these functions and a check for correctness
is desired.