Reinstate sequential implementation in case the model does not support asynchronous evaluations.
Implement parallel_flag and safeguard the parallel batch evaluation.
This is to avoid when the model itself cannot support run in parallel but batch_size is requested.
To handle this, we throw a warning, proceed in sequential mode, and ignore the batch_size requested.
Preliminary tests indicate that at least all tests are passed, including the warning. More tests are needed.