Properly apply scales and weights to field primary responses
DP-171: Minimal change bug fix to address ambiguity in how scales are
expressed and applied for scalars vs. fields. This will allow primary
scales and weights to be specified on a per response group basis and
apply to elements of fields.

NOTES: There is considerable inefficiency in this implementation to be
addressed in subsequent revisions.