DP-128:  Fix MSVC compile error resulting from branch merge
The windows compiler requires a return value if a non-void return
is specified in the function prototype.  Corrected the error by
returning 0 at the end of a function.

Request review by Brian or Mike since it appears that my fix may
NOT the correct way to address the issue.  Perhaps the function
should return void??