ML FT: finish debugging ML/MF FT with UNIFORM_MAX_RANK refinement.
UNIFORM_MAX_RANK is currently a hard-wired option for generating uniform
refinement candidates.  As opposed to incrementing the start rank or start
order, this option increments the max rank attainable within the adapt_rank
procedure and sizes the sample profiles based on the max regression size.

As long as the recovered ranks (from cross-validation) are active on the max
bound, refinement candidates continue to be generated by advancing the max rank
and incrementing the samples based on the new max regression size.  Once the
recovered ranks all fall below the current maximum, the model level has
saturated and does not generate any additional candidates for use in either
individual refinement or integrated greedy refinement.

Other details:
> Revise how initialization vs. sequence updates are performed from
  NonD*FunctionTrain by distinguishing SharedC3ApproxData::set_parameter()
  from set_active_parameter().  Add maxRank and randomSeed to active params.
> Replace hard-wired seed control with user seed for managing repeatability
  in C3 regression solves.  This requires a C3 patch that will be pushed once
  devel is reopened.
> TO DO's:
  >> determine whether the type of uniform refinement should become a user
     option (or can be down-selected)
  >> resolve some significant differences between our 1-year old version of
     the C3 library and Alex's current python implementation.