Surrogates: reduced order polynomial for GP
DP-95: Updated gold values in "unit test" by @rhoope.

Added a "none" keyword to the trend part of the input spec for
the GP that enables regression without a trend function.

When the trend keyword is omitted, a reduced quadratic trend is
used, consistent with what is done in the Surfpack GP. We may want
to change this default setting to "none" in the future.