surrogates: variance eval from Dakota's GP wrapper
DP-226: Implemented a prediction_variance method in DakotaSurrogatesGP
for use in EGO and related algorithms.

Added a value method to SurrogatesBase and its children that evaluates
the surrogate a single point (as opposed to a matrix of samples). This
method is used in Dakota's interfaces to the surrogates modules. Also
added a method for computing GP variance at a single point in the
surrogates module and Dakota wrapper class.

A test has been added to dakota_cantilever_ego that uses the surrogates
module GP (enabled by the "gaussian_process experimental" keyword).
Results are comparable to tests that use the Dakota or Surfpack GP.

EGO with the surrogates module GP is slower than the other variants as the
size of the build space grows. This appears to be related to an
unexpectedly slow linear solve in the negative marginal log-liklehood
evaluation.

Defined a floating point difference tolerance "near_zero" in
util_data_types.hpp that is used in the DataScaler and GPObjective
classes.