Make Results objects assignable
This commit adds __setitem__ to the BatchResults class, so that
Results objects can be assigned. This permits, e.g., Results
objects that are returned by functions to be added to a batch.

e.g.

  bp, br = di.read_parameters_file(batch=True)
  br[0] = calculate(bp[0], br[0])

The Results objects themselves received a new method, update(),
which takes the Results object to be copied in and the boolean
keyword argument 'check'. When check = True, several properties
of the incoming Results object are confirmed to match the
destination one, and a ResultsUpdateError is raised if there
are any differences. By default, check = False. So, more
carefully written code might look like:

  bp, br = di.read_parameters_file(batch=True)
  try:
      br[0].update(calculate(bp[0], br[0]), check=True)
  except di.ResultsUpdateError:
      print("calculate borked the Results object!")

BatchResults.__setitem__ calls update() with check = False.