Bayes opt: resolve a few remaining valgrind issues.
Previous commit corrected hidden convergenceTol, resulting in an uninitialized
memory error.  This commit resolves issues with finalize_run() for NCSUOptimizer
use within GaussProcApproximationm (lacked protection for a null iteratedModel).

Also added support for {initialize,finalize}_mapping() on approxSubProbModel in
EffGlobalMinimizer (as done in expansion and reliability methods that create
local Model recursions), while tracking cause of uninitialized memory.  This
did not turn out to be the cause of the memory errors, but is still needed to
support general Model recursions.