Surrogates: GP speed refactor
DP-226: Rewrote sections of the GP and GPObjective classes
to cache Gram matrix factorization and linear solve results
in the negative marginal log-likelihood (nmll) function when
possible.

Previous behavior in the nmll was to compute objective function
and gradient. Now there is a flag for gradient computation so that
it is only computed when needed by ROL. This results in some
computational savings because there is a O(N^3) linear solve in the
gradient calcuation.

Also loosened tolerances and max number of line search function
evaluations in ROL settings. There was no need to update any unit
tests, but the baseline for dakota_cantilever_ego that uses the
experimental GP needed to be updated.