DP-260: Count other kinds of UnitTest failures
Number of unittest passes and failures reported in the
nightly email are based on summary output computed
by a function in this file.

The counts were based on counting instances of the
strings "Passed" and "Failed" in CTest console output
for the unittests. However, CTest does not report all
kinds of failures using the "Failed" string, so this method
sometimes was inaccurate.

Switched to parsing a summary string that appears toward the
end of the CTest output. It appears to be more reliable.