Surrogates: refactor value and variance calls
DP-249 and DP-250: Changed the value and variance functions (only used
in EGO at for now) to return a VectorXd instead of a double, as this is
more general and friendly to future extensions. These calls now expect a
matrix of prediction points instead of a row vector for a single point.

There are now two signatures for the value and variances calls instead
of a single one with a default argument to make life easier on the
Pybind11 side.

This setup may be extended to gradient and Hessian calls if it passes
review. Should also consider changing the void value call for a matrix
of prediction points and all QoIs to a MatrixXd return type or remove
it altogether.