More portable Python-based regression tests
Configure dakota_batch and dakota_workdir tests so that the path to
the Python executable is embedded in the specified
analysis_driver. This allows testing to use the Python Dakota was
built with instead of a system Python.

Done in a way to preserve running dakota_test.perl in
build/test. Initial pass will break installation testing of Dakota by
end users for these two tests.

This commit also includes some other CMake extensions and cleanup:

 * Always copy all input and associated aux files into build/test

 * Assume build tree test files are in pwd, but still assume baselines
   are in CMAKE_CURRENT_SOURCE_DIR. This allows testing an input file
   without reference to the source tree.

 * Extend SEMS AddApplicationTest with a CONFIGURE_FILES option

 * Cleanup redundant file copies between serial and parallel tests