Specify Python interp in Windows workdir test
Specify the path to Dakota's Python for dakota_workdir_windows.in.
Update the script to work with Python 3 and import print_function in
both Unix and Windows variants as a caution.

NOTE: Left Windows system tests using a bare .py analysis_driver, while
fork cases specify the Python in this initial cut.