Add input spec for choosing MLMC optimization formulation.
Add input spec convergence_tolerance_target which can be set to variance_constraint (default) or cost_constraint. For variance_constraint we fine the MLMC sample allocation by minimizing the cost for an equality constraint on the variance. For cost_constraint it is the other way around, i.e., minimizing the variance for an equality constraint on the cost. This setting also has an effect for convergence_tolerance_type. Namely, when set to relative (default) we find the sample allocation with respect to the cost of the given number of evaluations on the fine grid. When set to absolute, we find the allocation with respect to that absolute value.