Refine ProblemDescDB lookup maps using member fn
DP-391: Remove lookup class in favor of templated member function. In
the course of debugging Windows builds for DP-400, Greg pointed out
that there's really no reason for the class.

This reduces object file size for ProblemDescDB.cpp.o by about 25% on
Linux and might be enough to resolve bloat on Windows that's resulting
in build failures.