bayes_opt: improve best sample extraction and finalization for batch-asynch EGO
Previous get_best_sample() was relying on fresh fHatModel evaluations (which
was a waste but not too big of a sin assuming GP interpolation), but now
extracts from the truth build data.

Upon convergence in batch-asynch EGO, complete any running evaluations and
include the new truth evals in the final post-processing for the best solution.

Added more low-level helpers to consolidate code.