DP-475: Refine two demos of adjusting Mac Boost dylibs
1. Finish the demo of embedding @rpath in various executables (note
it's also necessary to patch Surfpack use of Boost::serialization for
this approach).

2. Add a new approach that addresses the need to fixup the Boost
libraries themselves to be able to find each other. This also
alleviates the need to embed any RPATH info.

If successful, (2) is an approach that can be set up by the driving
Jenkins job and need not bleed into the core Dakota CMake.