DP-499: Write configOptions to Yaml during save
Surrogates now write their configOptions ParameterList to a Yaml
file when they are serialized so that a record of these options
exists.

The GP has a few Eigen containers in its configOptions that show
up as blanks in the written Yaml file and this may be remedied in
a future commit.