Bug: Input validation given both fields and constraints
DP-497: Fix bug where a mixture of primary fields (calibration terms)
and nonlinear constraints resulted in an incorrect error about
response apportionment between scalars and fields. Manifested itself
as error about unexpected number of descriptors.

Also cleaned up a number of related issues with the construction of
ExperimentData and associated ExperimentResponse, dimensions of
covariance, etc. All had to be updated to ensure that
experiment-related data is sized to number of primary responses, not
the total number of functions.