non_hierarch: fix CV range for ML FTT with RANK_SAMPLING
NonDMultilevelFunctionTrain::assign_allocation_control() turns on adapt_rank
for the case of ML FTT with sample allocations based on the RANK_SAMPLING
option.  The latest code was not tolerant of default max_rank of SZ_MAX or
default max_order of USHRT_MAX, and this commit now accommodates these cases.