ExperimentData: Keep class in sync on data adds
Refine the hack I provided tporton for cd2c0b9f. When ExperimentData
is constructed or updated on-the-fly (as in Hi2Lo), make sure related
class members for lengths/offsets and covariance determinant get
updated.

Also correct a long-standing TODO so that all stored Responses are
indeed ExperimentResponses.

There's clear room for optimization, but this should clean things up
and make it more self-consistent.