Cleanup use of Eigen, including allowing external
DP-479: Quietly probe for Eigen3 version 3.3 or later using typical
CMake means first, and if not found, fallback on Dakota's. If this
logic doesn't suffice we can add overrides later. Only install
Dakota's Eigen if using it.

Consolidate various separate calls to find_package for Eigen into a
single place, and use the results throughout, including for passing to
MUQ and NLOPT.

Remove conditionals as we always include Eigen now.