non_hierarch: add rudimentary control by maxFunctionEvals
Just an additional convergence check on equivHFEvals for each while loop,
with no additional inter-iteration granularity.  It is generally possible to
tailor the sample profile to a prescribed budget, but not supporting that on
this initial pass.

TO DO: revisit (minimizer-based) defaults for maxFunctionEvals.  Should we
just use SZ_MAX throughout?  While we don't want to allow run-aways, both
maxIterations and maxFunctionEvals defaults can truncate runs unexpectedly...