Lawson-Hanson: Fix arrays for GCC 10 compilation
DP-623: Fix Lawson-Hanson Fortran warnings turned errors with GCC 10
where scalars were being passed to procedures requiring arrays.

"A procedure expected an array but a scalar was passed.  Since Fortran
passes everything by reference, it works out and is pretty common in
practice.  I simply declared an array of proper length, copied the
scalar passed in to the array, and sent the array."

Thanks to Tim Fuller for the patch and Laura Swiler for reviewing/testing!

Also update SHA1 for packages/surfpack for similar MARS fixes.