Improve handling of types
dakota.interfacing.read_parameters_file() has two new keyword
parameters.

infer_types: Automatically guess types. True by default. This is
    the legacy behavior. When False (and no types are specified)
    all the types will be str.
types:  list or dict of type overrides. If a list, must have a
    number of elements equal to the number of variables. If
    a dict, then variables whose descriptors match keys will
    have their types overriden. For both list and dict, the
    types are expected to be single argument functions.
    Currently, this is not verified, and if violated an ugly
    exception will result. If infer_types = True and types
    are provided, the latter take priority.

In addition, the type inference was improved a bit to check
whether a dakota variable contains an underscore. If so, it is
assumed to be a string.