Response metadata: Add metadata requests to param files
DP-755: Include number of metadata and requested field names to the
list of data requests made in a parameters file. Specific format may
still evolve.

Minimal dakota.interfacing updates to free tests. This adds parsers
for the metadata entries in the files, but ignores them for now. Need
to add to the Results class, most likely.