Persist response metadata across numerical derivatives
DP-769: An attempt to "swallow" metadata at the Model level when
estimating finite difference derivatives doesn't work because the same
Response object is used to both retrieve the underlying evaluations
and aggregate the final Response with derivatives. Needs a redesign
for this use case.

Side-effect: for now, Models may allocate, but not populate metadata
in finite difference cases.