Make hdf5 tests compatible with h5py 3.0+
h5py 3.0 and forward treats the data in string datasets as
byte strings. To access Python strings, it's necessary to use
the asstr() accessor. This change creates a couple of helper
functions to smooth over this change and make the tests work
with h5py versions newer and earlier than 3.0.