Fix hdf5 lib packaging issue
Resolve problem discovered by EMR -- that some libraries were
not being installed into bin/ in our packages.

The problem: When packaging or installing, we copy TPL
dependencies into bin/. To figure out which libraries to copy, we
run otool -L on all the targets that CMake installs into bin/
and lib/ (and a few other places). We exclude entries that are
obvious system paths (presumably all users have those already)
or that begin with @rpath, which we assumed indicates that they
were built by CMake.

Unfortunately, with the transition to the new TPL system, some of
the TPLs now also have ids that include @rpath, and were being
skipped.

We hope to eventually switch to use a new CMake feature,
install(RUNTIME_DEPENDENCY_SET), but it requires 3.21 and was
judged too big of a change so close to a release.

The workaround is to use another CMake feature,
file(RUNTIME_DEPENDENCIES), to scan a few executables in the
build tree to get a more complete picture of which TPLs need
to be copied into bin.