ensemble_select: modify analytic and reordered analytic solves to use averaged evaluation ratios
This makes analytic approaches more consistent with the numerical solve
formulations for MFMC, ACV, and GenACV (which employ QoI-averaged counts or
eval ratios as design variables).  The key driver for making this change is
that the estimator variance calculations should not take credit for varying
the eval_ratios per QoI when shared sampling increments employ averages.  That
is, estimator variance calculations should honor the shared sampling constraint.

TO DOs:
> additional testing of consistency among analytic MFMC, numerical MFMC, GenACV
> deploy approach to NonDMultilevControlVarSampling for consistency there