DP-1096: Fix correlation output labels
The order of variable labels for correlation matrix output was
not consistent with the numerical results when more than one
category (e.g. design, aleatory) of variable were present in
a study. Numerical results were in "input spec" order, and
labels were in order of cv, div, dsv, drv. This was the case
for both console output and HDF5.

This change includes:
- New methods on Model to return the "input spec" ordered labels.
  These use the Variables::ordered_labels() method that BMA added
  a couple of years ago.
- Updates to SensAnalysisGlobal::print/archive_correlations to
  take a single list (StringArray) of variables to use as labels.
- Updates to methods that write and archive correlations to use
  the new Model::ordered_labels() and updated SensAnalysisGlobal
  methods.
- Code clean up of SensAnalysisGlobal
- New "unit" test that verify the console and HDF5 output for a
  mixed variable category study.
- A verification test of the correlations.
- Updated baselines for affected regression tests