Fix links to keywords and topics
Automatic translation of the old keyword files created
hyperlinks to other keywords and topics that are a bit
unaesthetic. E.g. a link to the interface-failure_capture
keyword was written as

:ref:`interface-failture_capture<interface_capture>`.

This commit attempts to update keyword hyperlinks to use
the :dakkw: role. The update was scripted, and keyword
hyperlinks were distinguished from other using two
criteria:

1. The target and link text had to match.
2. The link text had to be the name of a keyword file.

I also updated the topic hyperlinks using similar critera.