Fix hdf5 moment labeling problem
My guess about why this change is necessary: The
StringScale does not make copies of the sequence of
labels passed into it. It assumes that the
objects passed in continue to exist for at least as
long it takes for them to be written to HDF5.

Previously the initializer_list constructor for the
StringScale was being used when NonDExpansion stored
moments. Here's the guess: Copies of the strings were
being created when the constructor was called and then
destroyed soon thereafter, perhaps at the constructor's
exit. As a result, the const char * pointers to the
strings data were not guaranteed to be valid and in some
cases crap was being written to HDF5.

On RHEL7 and gcc/4.8.5, we were "luckily" not seeing
this problem, but newer gcc and clang just worked
differently.

Until I wrote a test, this issue was just getting by us.
A user complained but I couldn't duplicate the problem and
didn't understand at all what was happening.