Safely create StringScales with init_lists of strings
Instead of working around the problem in NonDExpansion,
add a StringScale constructor that does the right thing.

What appears was happening is that in the absence of
an explicit std::initializer_list<String> constructor,
the initializer_list was being "cast" to vector<String>,
and *that* constructor was being called.

But somewhere in that conversion, the strings were being
copied (new objects created). Those copies were being
destructed at the exit of the StringScale constructor,
which invalidated the pointers in StringScale::items.

The sol'n, which is clunky, is to add a class member
that caches the strings only for the initializer_list<String>
constructor. My goal for these *Scale classes was
to avoid making copies, but the objects typically will
be short-lived, and, because they use an initializer_list,
almost certainly will be small.