Fix bugs uncovered by recent PR
The PR corrected an amateurish Python mistake I made long ago.

Before that, this test was giving false positives; nothing was
being tested. The change uncovered bugs in the test script
and input file itself. Fortunately, after fixing them, the
features of Dakota it was supposed to be protecting appear
to be fine.