Close files when managing objects are destructed
When Dakota is used in "library mode", such as with
the environment Python module, destruction of the
LibraryEnvironment doesn't clean everything up. Some objects
and other resources have a global scope and are only
cleaned up at process termination. With this commit,
JEGA's global log and files that NPSOL writes
to are closed upon destruction of the Dakota objects
that interact with JEGA and NPSOL.