Add support for JSON parameters and results files
This commit introduces two new interface keyword
groups: parameters_format and results_format. The
aprepro option is moved under parameters_format,
and standard (the default) and json formats are
also available.

On the results side, there are two options,
standard (the default) and json. The labeled
keyword is now a sub-keyword of standard.

The commit also adds a new header-only TPL
to Dakota to support JSON.