ensemble_select: adding option to recur on hierarchical DAGs, hardening for MFMC/MLMC promotions
MFMC and MLMC "promotions" to GenACV now allow for recursion across a full
set of hierarchical DAGs (essentially reordering the hierarchy), in addition
to model selection.

Also trapping cases where computed eval ratios could be < 1 for analytic
allocations, corrupting the R^2 calculation.  This mirrors the bounds and
constraints in the numerical cases, and has been observed to occur for the
transient diffusion problem, which collapses a 2D hierachy head to tail.