Make uSpaceModel a pointer in NonDExpansion
The first Model member function I began trying to shift down
to ProbabilityTransformModel is probability_transformation().

One place this member function is used (implicitly) is in
the NonDExpansion hierarchy. In most derived classes,
iteratedModel is wrapped in a ProbabilityTransformModel,
which is then used to construct uSpaceModel, which is a
DataFitSurrogate.

This is a case where a SurrogateModel and RecastModel both
override member functions that properly belong to other one
so that recast models can be surrogated, or surrogates can be
recast.

So, even though my goal was to work on ProbabilityTransformModel,
I instead wound up working on SurrogateModel, especially in
the context of NonDExpansion.

I'm guessing many/most of the Recast and Surrogate-related
member functions will have these interactions.

The change here is to change uSpaceModel from a Model variable
to a shared_ptr<DataFitSurrModel>. It's a shared_ptr instead of
unique_ptr because of something a little messy in
the NonDSurrogateExpansion constructor. If it weren't for that,
a unique_ptr would have been fine for all other derived classes.

It does need to be some kind of smart pointer because most of
the derived classes plainly manage its lifetime.