Check for default constructed iterator
This commit fixes a segfault caused by a couple of
interacting changes previously made.

After making subIterator a shared_ptr<Iterator>, I naively
converted a conditional that tests subIterator.is_null() to
check whether subIterator is itself set. This results in a
segfault in NestedModel when Dakota is run in parallel
due changes in 1769ef09.

That commit "default constructs" Iterators in
IteratorScheduler::init__iterator() on non-zero ranks when
Dakota is run in parallel in order to cache information
in them that originally (on devel) was being stored in
Iterator envelopes.

Because the envelopes had no letter, the conditional
on is_null() was bypassed, but now is executed because
subIterator is no longer null.

This change tests subIterator->method_id(), assuming that
when it is empty, subIterator points to a default
constructed object.

I would very much welcome comments/suggestions on this
approach. It feels pretty non-ideal.

An alternative I considered was adding a new member variable
and getter to Iterator, Iterator::default_constructed(). But
I hesitated because I didn't want to add a whole new member
function for this one unusual situation.