Fix Windows build errors
MSVS treats non-void functions that do not
return anything to be an error. There are a
handful of these in Iterator, where virtual
functions that have no reasonable default
behavior call the abort handler and return
nothing. A remedy I discovered while
working on Model is to unconditionaly throw
an exception. The throw is never reached
because of the call to abort, but it satisifies
the compiler that the lack of return is okay.