Flip over to hardened solutions based on truncated SVD throughout non-hierarchical estimators
Following up on the Teuchos SDP bug, the Cholesky approach was not doing as
well as previously thought.  Truncated SVD shows improvement in almost all
cases, even though it can give up a bit of accuracy due to truncation.

Blessing the easiest cases, but have a bit more to do.