Fix study lookup
Earlier version used the address of the envelope instead of
the rep. To use the rep, I added a member to PDB that returns
it.

Testing also uncovered a case where failing to clear the
iterator and model caches after completion of a study caused
a problem. I planned to do this, but hadn't hooked it up. I
added a function in a separate file to decouple Iterator, Model,
from ProblemDescDB.